+++ NEWS.md
@@ -1,5 +1,9 @@
 # teal.goshawk 0.1.11.9006
 
+### Breaking Changes
+* Converted the `hline` parameter of `tm_g_gh_lineplot` to three parameters: `hline_arb`, `hline_arb_color` and `hline_arb_label`.
+
+### Miscellaneous
 * Added basic logging to the modules.
 * Rewrote modules to use `moduleServer` and updated calls to `teal.devel` modules which have also been written to use `moduleServer`.
 * Replaced calls to `teal::root_modules` with `teal::modules` following deprecation of `teal::root_modules`.